+++ frontend/src/lib/LayoutView/LayoutView.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+  import { setState } from '$src/globalStates.svelte';
+  import i18n from '$src/i18n.svelte';
   import DocumentOptions from './DocumentOptions.svelte';
   import SetEditorPage from './SetEditorPage.svelte';
   import SkeletonPage from './SkeletonPage.svelte';
@@ -6,17 +8,26 @@
   let activePage = $state<'skeleton' | 'options'>('skeleton');
 </script>
 
-<div class="layout-container">
-  <SkeletonPage isActive={activePage === 'skeleton'} />
-  <SetEditorPage isActive={activePage === 'options'} />
-</div>
-<div class="footer switch-footer">
-  <button onclick={() => (activePage = 'skeleton')}>Layout</button>
-  <button onclick={() => (activePage = 'options')}>Options</button>
-</div>
-<div class="footer options-footer">
-  <DocumentOptions />
-</div>
+{#if setState.addedSets.length == 0}
+  <div class="text-container">
+    <h2>{i18n.t('layout_no_set_added')}</h2>
+    <p>
+      {i18n.t('layout_add_set_instruction')}
+    </p>
+  </div>
+{:else}
+  <div class="layout-container">
+    <SkeletonPage isActive={activePage === 'skeleton'} />
+    <SetEditorPage isActive={activePage === 'options'} />
+  </div>
+  <div class="footer switch-footer">
+    <button onclick={() => (activePage = 'skeleton')}>Layout</button>
+    <button onclick={() => (activePage = 'options')}>Options</button>
+  </div>
+  <div class="footer options-footer">
+    <DocumentOptions />
+  </div>
+{/if}
 
 <style>
   .layout-container {
@@ -24,6 +35,7 @@
     flex: 1;
     display: flex;
     width: 100%;
+    justify-content: center;
     min-height: 0;
     overflow-y: auto;
   }
@@ -52,6 +64,14 @@
     button {
       width: 12rem;
     }
+  }
+
+  .text-container {
+    display: flex;
+    height: 100dvh;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
   }
 
   @container main (width < 70rem) {