+++ frontend/src/lib/LayoutView/SkeletonPage.svelte
@@ -1,44 +1,59 @@
 <script lang="ts">
+  import { adjustValue } from '$src/commonFunctions.svelte';
   import { documentOptions, setState } from '$src/globalStates.svelte';
-  import type { SetState } from '$src/types';
   import SetSkeleton from './SetSkeleton.svelte';
+  import i18n from '$src/i18n.svelte';
+  import AnswerSkeleton from './AnswerSkeleton.svelte';
 
   let { isActive }: { isActive: boolean } = $props();
 
   let pages = $derived.by(() => {
-    let pages: SetState[][] = [[]];
-    setState.addedSets.forEach(set => {
-      pages[pages.length - 1].push(set);
-      if (set.options.set_options.pagebreakAfter) {
+    let pages: number[][] = [[]];
+    for (let i = 0; i < setState.addedSets.length; i++) {
+      pages[pages.length - 1].push(i);
+      if (setState.addedSets[i].options.set_options.pagebreakAfter) {
         pages.push([]);
       }
-    });
+    }
     // Remove dangling page if last set has pagebreak
     if (pages[pages.length - 1].length === 0) {
       pages.pop();
     }
     return pages;
   });
+  let xMargin = $derived(adjustValue(0, 40, documentOptions.xMargin, 0, 2, 1));
+  let yMargin = $derived(adjustValue(0, 40, documentOptions.yMargin, 0, 2, 1));
 </script>
 
 <div class="page-container {isActive ? 'open' : 'closed'}">
-  {#each pages as page, i}
-    <div class="page">
-      {#if i === 0}
-        <h1>First page</h1>
-      {/if}
-      {#each page as set}
-        <SetSkeleton {set} />
-      {/each}
-      {#if i === pages.length - 1 && !documentOptions.pageBreakBeforeAnswers}
-        <h1>Answers</h1>
-      {/if}
-    </div>
-  {/each}
-  {#if documentOptions.pageBreakBeforeAnswers}
-    <div class="page">
-      <h1>Answers</h1>
-    </div>
+  {#if setState.addedSets.length == 0}{:else}
+    {#each pages as page, i}
+      <div
+        class="page"
+        style="--x-margin: {xMargin}rem; --y-margin: {yMargin}rem;"
+      >
+        {#if i === 0}
+          <input
+            name="title"
+            class="input title"
+            placeholder={i18n.t('title_placeholder')}
+            type="text"
+            bind:value={documentOptions.title}
+          />
+        {/if}
+        {#each page as index, j}
+          <SetSkeleton bind:set={setState.addedSets[index]} index={j} />
+        {/each}
+        {#if i === pages.length - 1 && !documentOptions.pageBreakBeforeAnswers}
+          <AnswerSkeleton ownPage={false} />
+        {/if}
+      </div>
+    {/each}
+    {#if documentOptions.pageBreakBeforeAnswers}
+      <div class="page">
+        <AnswerSkeleton ownPage={true} />
+      </div>
+    {/if}
   {/if}
 </div>
 
@@ -47,23 +62,29 @@
     flex: 1 1 auto;
     width: 100%;
     min-height: 0;
+    height: fit-content;
     min-width: 0;
+    max-width: 52rem;
 
     display: flex;
     flex-direction: column;
-    gap: 0.5rem;
+    gap: 1.5rem;
 
     background: none;
     padding: 1rem;
   }
 
   .page {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
     width: 100%;
     flex-shrink: 1;
-    padding: 0.5rem;
+    padding: var(--y-margin) var(--x-margin);
     background-color: var(--bg-light);
     border: 1px solid var(--border);
     box-shadow: var(--shadow-elevation-low);
+    transition: all 0.4s;
   }
 
   @container main (width < 70rem) {
@@ -75,9 +96,17 @@
     }
   }
 
-  h1 {
-    font-size: 2rem;
+  .input {
+    background: none;
+    border: none;
     text-align: center;
+    field-sizing: content;
+    width: fit-content;
+  }
+
+  .title {
+    margin-top: 1.5rem;
     margin-bottom: 2rem;
+    font-size: 1.2rem;
   }
 </style>